몸무게 double로 변경 / 챌린지 완료되는 상태 -> 등록당시 증가,감소에 따라 조건 수정

+++ src/main/java/com/ai/balancelab_be/domain/challenge/dto/ChallengeDTO.java
@@ -18,9 +18,9 @@ public class ChallengeDTO {
 
     private String periodUnit;
 
-    private Integer startWeight;
+    private double startWeight;
 
-    private Integer targetWeight;
+    private double targetWeight;
 
     private LocalDate startDate;
